mkfs.fat: Avoid usage of undefined behavior in open()
Behavior of O_EXCL in open() is undefined if it is used without O_CREAT.
The only exception on Linux is when pathname is block device. In that case
when block device is already mounted, open() fails with error EBUSY.

This change avoids calling O_EXCL on path names which are not block
devices. First it opens path name normally, check file descriptor that it
is block device, and then re-open file descriptor with O_EXCL via /proc.
This is fully race-free trick without any undefined behavior in usage.